+++ R/manip.r
@@ -379,12 +379,26 @@ arrange.grouped_df <- function(.data, ..., .by_group = FALSE) {
 #' accidental matching of data frame variables when you refer to
 #' variables from the calling context.
 #'
+#'
 #' @section Scoped selection and renaming:
 #'
 #' The three [scoped] variants of `select()` ([select_all()],
 #' [select_if()] and [select_at()]) and the three variants of
 #' `rename()` ([rename_all()], [rename_if()], [rename_at()]) make it
 #' easy to apply a renaming function to a selection of variables.
+#'
+#'
+#' @section Context of evaluation:
+#'
+#' Quoting verbs usually support references to both objects from the
+#' data frame and objects from the calling context. Selecting verbs
+#' behave a bit differently.
+#'
+#' * Bare names are evaluated in the data frame only. You cannot refer
+#'   to local objects unless you explicitly unquote them with `!!`.
+#'
+#' * Calls to helper functions are evaluated in the calling context
+#'   only. You can safely and directly refer to local objects.
 #'
 #' @inheritParams filter
 #' @inheritSection filter Tidy data
@@ -407,6 +421,7 @@ arrange.grouped_df <- function(.data, ..., .by_group = FALSE) {
 #' @export
 #' @examples
 #' iris <- as_tibble(iris) # so it prints a little nicer
+#'
 #' select(iris, starts_with("Petal"))
 #' select(iris, ends_with("Width"))
 #'
@@ -420,6 +435,21 @@ arrange.grouped_df <- function(.data, ..., .by_group = FALSE) {
 #'
 #' # Drop variables with -
 #' select(iris, -starts_with("Petal"))
+#'
+#'
+#' # When selecting with bare symbols, you can only refer to data
+#' # frame objects. This avoids ambiguity. If you want to refer to
+#' # local objects, you can explicitly unquote them. They must contain
+#' # column positions (integers) or column names (strings):
+#' Species <- 2
+#' select(iris, Species)     # Picks up `Species` from the data frame
+#' select(iris, !! Species)  # Picks up the local object referring to column 2
+#'
+#' # On the other hand, function calls behave the opposite way. They
+#' # are evaluated in the local context only and cannot refer to data
+#' # frame objects. This makes it easy to refer to local variables:
+#' x <- "Petal"
+#' select(iris, starts_with(x))  # Picks up the local variable `x`
 #'
 #'
 #' # The .data pronoun is available: