+++ tests/testthat/test-select.r
@@ -51,26 +51,6 @@ test_that("negating empty match returns everything", {
 
 # Select variables -----------------------------------------------
 
-test_that("select_vars can rename variables", {
-  vars <- c("a", "b")
-  expect_equal(select_vars(vars, b = a, a = b), c("b" = "a", "a" = "b"))
-})
-
-test_that("last rename wins", {
-  vars <- c("a", "b")
-
-  expect_equal(select_vars(vars, b = a, c = a), c("c" = "a"))
-})
-
-test_that("negative index removes values", {
-  vars <- letters[1:3]
-
-  expect_equal(select_vars(vars, -c), c(a = "a", b = "b"))
-  expect_equal(select_vars(vars, a:c, -c), c(a = "a", b = "b"))
-  expect_equal(select_vars(vars, a, b, c, -c), c(a = "a", b = "b"))
-  expect_equal(select_vars(vars, -c, a, b), c(a = "a", b = "b"))
-})
-
 test_that("select can be before group_by (#309)", {
   df <- data.frame(
     id = c(1, 1, 2, 2, 2, 3, 3, 4, 4, 5),
@@ -99,77 +79,8 @@ test_that("rename does not crash with invalid grouped data frame (#640)", {
   )
 })
 
-test_that("can select with character vectors", {
-  expect_identical(select_vars(letters, "b", !! "z", c("b", "c")), set_names(c("b", "z", "c")))
-})
-
-test_that("abort on unknown columns", {
-  expect_error(select_vars(letters, "foo"), "must match column names")
-  expect_error(select_vars(letters, c("a", "bar", "foo", "d")), "bar, foo")
-})
-
 test_that("rename() handles data pronoun", {
   expect_identical(rename(tibble(x = 1), y = .data$x), tibble(y = 1))
-})
-
-
-# combine_vars ------------------------------------------------------------
-# This is the low C++ function which works on integer indices
-
-test_that("empty index gives empty output", {
-  vars <- combine_vars(letters, list())
-  expect_equal(length(vars), 0)
-
-  vars <- combine_vars(letters, list(numeric()))
-  expect_equal(length(vars), 0)
-})
-
-test_that("positive indexes kept", {
-  expect_equal(combine_vars(letters, list(1)), c(a = 1))
-  expect_equal(combine_vars(letters, list(1, 26)), c(a = 1, z = 26))
-  expect_equal(combine_vars(letters, list(c(1, 26))), c(a = 1, z = 26))
-})
-
-test_that("indexes returned in order they appear", {
-  expect_equal(combine_vars(letters, list(26, 1)), c(z = 26, a = 1))
-})
-
-
-test_that("negative index in first position includes all others", {
-  vars <- combine_vars(letters[1:3], list(-1))
-  expect_equal(vars, c(b = 2, c = 3))
-})
-
-test_that("named inputs rename outputs", {
-  expect_equal(combine_vars(letters[1:3], list(d = 1)), c(d = 1))
-  expect_equal(combine_vars(letters[1:3], list(c(d = 1))), c(d = 1))
-})
-
-test_that("if multiple names, last kept", {
-  expect_equal(combine_vars(letters[1:3], list(d = 1, e = 1)), c(e = 1))
-  expect_equal(combine_vars(letters[1:3], list(c(d = 1, e = 1))), c(e = 1))
-})
-
-test_that("if one name for multiple vars, use integer index", {
-  expect_equal(combine_vars(letters[1:3], list(x = 1:3)), c(x1 = 1, x2 = 2, x3 = 3))
-})
-
-test_that("invalid inputs raise error", {
-  expect_error(
-    combine_vars(names(mtcars), list(0)),
-    "Each argument must yield either positive or negative integers",
-    fixed = TRUE
-  )
-  expect_error(
-    combine_vars(names(mtcars), list(c(-1, 1))),
-    "Each argument must yield either positive or negative integers",
-    fixed = TRUE
-  )
-  expect_error(
-    combine_vars(names(mtcars), list(12)),
-    "Position must be between 0 and n",
-    fixed = TRUE
-  )
 })
 
 test_that("select succeeds in presence of raw columns (#1803)", {
@@ -179,7 +90,7 @@ test_that("select succeeds in presence of raw columns (#1803)", {
   expect_identical(select(df, -b), df["a"])
 })
 
-test_that("arguments to select() don't match select_vars() arguments", {
+test_that("arguments to select() don't match vars_select() arguments", {
   df <- tibble(a = 1)
   expect_identical(select(df, var = a), tibble(var = 1))
   expect_identical(select(group_by(df, a), var = a), group_by(tibble(var = 1), var))
@@ -189,7 +100,7 @@ test_that("arguments to select() don't match select_vars() arguments", {
   expect_identical(select(group_by(df, a), include = a), group_by(tibble(include = 1), include))
 })
 
-test_that("arguments to rename() don't match rename_vars() arguments (#2861)", {
+test_that("arguments to rename() don't match vars_rename() arguments (#2861)", {
   df <- tibble(a = 1)
   expect_identical(rename(df, var = a), tibble(var = 1))
   expect_identical(rename(group_by(df, a), var = a), group_by(tibble(var = 1), var))
@@ -203,4 +114,16 @@ test_that("can select() with .data pronoun (#2715)", {
 
 test_that("can select() with character vectors", {
   expect_identical(select(mtcars, "cyl", !! "disp", c("cyl", "am", "drat")), mtcars[c("cyl", "disp", "am", "drat")])
+})
+
+test_that("select() evaluates symbols outside of context", {
+  foo <- 2
+  expect_error(select(mtcars, foo), "object 'foo' not found")
+})
+
+test_that("select() supports parenthesised expressions", {
+  foo <- 2
+  expect_error(select(mtcars, ((foo))), "object 'foo' not found")
+  expect_identical(select(mtcars, -((((cyl)):((carb))))), select(mtcars, mpg))
+  expect_identical(select(mtcars, -((contains("yl")))), select(mtcars, -cyl))
 })